refactor(update): point consumers at components/AppUpdate + mount AppUpdateForeground
Follow-up to the previous commit's UpdateReminder/hooks.tsx → AppUpdate/*
split. Without this commit, the prior commit's deletion of hooks.tsx
leaves 11 import sites pointing at a missing module, so the two must
ship together.

- 11 import paths migrated from `components/UpdateReminder/hooks` to
  `components/AppUpdate` (consumers in MoreActionButton, UpdateReminder,
  useVersionCompatible, views/AppUpdate/*, views/CloudBackup,
  views/Setting/Tab/*).
- Bootstrap.tsx now mounts <AppUpdateForeground /> as a render-nothing
  child of <Bootstrap />. Replaces the per-mount first-launch dispatch
  and AppState listener that previously lived in useAppUpdateInfo's
  autoCheck branch — now exactly one global listener for the app's
  lifetime.
- Bootstrap's Desktop-only useAppUpdateInfoCallback drops the second
  `autoCheck` arg since side effects no longer live in the hook.
- apps/mobile/jest.harness.config.mjs: update the jsdom-environment
  ignore list to point at the renamed test paths
  (UpdateReminder/hooks.test.ts → AppUpdate/useAppUpdate.test.ts) and
  pre-list the planned AppUpdateForeground.test.tsx.
