+++ packages/kit/src/views/Setting/pages/Tab/CustomElement.tsx
@@ -42,7 +42,7 @@ import PasswordUpdateContainer from '@onekeyhq/kit/src/components/Password/conta
 import {
   isShowAppUpdateUIWhenUpdating,
   useAppUpdateInfo,
-} from '@onekeyhq/kit/src/components/UpdateReminder/hooks';
+} from '@onekeyhq/kit/src/components/AppUpdate';
 import useAppNavigation from '@onekeyhq/kit/src/hooks/useAppNavigation';
 import { TabFreezeOnBlurContext } from '@onekeyhq/kit/src/provider/Container/TabFreezeOnBlurContainer';
 import {
